:construction: Attempting to break wizard steps out into a composition function.
