+++ web/src/pages/my-travel-requests/details/DetailsSubmitPage.vue
@@ -58,9 +58,7 @@ async function validateAndSave() {
       return false
     }
 
-    await approvalsEditFormCard.value.save({
-      stepNumber: travelAuthorization.value.stepNumber + 1,
-    })
+    await approvalsEditFormCard.value.save()
     snack.success("Travel request saved.")
     emit("updated", travelAuthorization.value.id)
     return true
